Papers

The module RealTimeScheduling.Papers provides implementations of proposed algorithms (schedulers, schedulabity tests, etc.) from published papers. These are typically of less broad applicability than the things implemented directly in the RealTimeScheduling package, but still of interest to the broader real-time community.

If you use RealTimeScheduling.jl in your research, please consider contributing your code to this module!

Submodules

RealTimeScheduling.Papers — Module

Implementations of proposed algorithms (schedulers, schedulabity tests, etc.) from published papers. These are typically of less broad applicability than the things implemented directly in the RealTimeScheduling package, but still of interest to the broader real-time community.

If you use RealTimeScheduling.jl in your research, please consider contributing your code to this module!

RealTimeScheduling.Papers.JobClassLevelScheduling — Module

An implementation of the scheduler and schedulability test from the paper "Toward Practical Weakly Hard Real-Time Systems: A Job-Class-Level Scheduling Approach" by Choi, Kim, and Zhu. DOI: 10.1109/JIOT.2021.3058215