Weakly Hard Systems

RealTimeScheduling provides basic support for weakly hard constraints.

Warning

Support for weakly hard systems is still a work in progress, and the API is still very incomplete. For now, we only support the constraints themselves, as well as comparisons between them.

RealTimeScheduling.WeaklyHardConstraint — Type

Abstract parent type of all weakly hard constraints

The concrete subtypes of this, and many of the methods defined on them, are due to Bernat, Burns, and Llamosí, "Weakly Hard Real-Time Systems," IEEE Trans. Computers, Vol. 50, No. 4, April 2001.

RealTimeScheduling.MeetAny — Type

MeetAny{T}(meet::T, window::T)

Weakly hard constraint specifying that a task meets meet deadlines in any window of size window.

RealTimeScheduling.MissAny — Function

MissAny([T=Int, ]miss, window)

Construct a MeetAny{T} constraint that allows at most miss deadlines to be missed in any window of size window.

While constraints of this sort are often expressed as a different type in the literature, they are equivalent to MeetAny, so using a single type simplifies any code using weakly hard constraints.

RealTimeScheduling.MeetRow — Type

MeetRow{T}(meet::T, window::T)

Weakly hard constraint specifying that a task meets meet deadlines in a row in any window of size window.

RealTimeScheduling.MissRow — Type

MissRow{T}(miss::T)

Weakly hard constraint specifying that a task misses at most miss deadlines in a row.