Scheduling

RealTimeScheduling implements a simple yet flexible scheduling simulator supporting preemptive global job-level fixed priority scheduling of TaskSystem structs. It also supports generating plots of the resulting schedules using the Plots.jl package.

using RealTimeScheduling, Plots
T = TaskSystem([PeriodicImplicitTask(3, 2), PeriodicImplicitTask(3, 2), PeriodicImplicitTask(3, 2)])
s = schedule_gedf(T, 2, 30.)
plot(s)

RealTimeScheduling.schedule_global — Function

schedule_global(release!, T, m, endtime)

Simulate a preemptive global schedule of task system T on m processors to the specified endtime. When releasing a job, the function release!(job) is called, allowing arbitrary modifications to be made, enabling a wide variety of global schedulers to be implemented. Two common examples are provided: schedule_gfp and schedule_gedf.

The job provided to release!(job) defaults to being released as early as possible (i.e. at time 0 or one period after the task's last job), and has the relative deadline and cost specified by the task. The priority defaults to the task's index in T, with lower priority values being treated as higher priority by the scheduler.

RealTimeScheduling.schedule_gedf — Function

schedule_gedf(T, m, time)

Simulate a preemptive global earliest-deadline-first (GEDF) schedule of task system T on m processors for the specified time.

See also schedule_global for more general global scheduling.

RealTimeScheduling.schedule_gfp — Function

schedule_gfp(T, m, time)

Simulate a preemptive global fixed priority (GFP) schedule of task system T on m processors for the specified time. Tasks are prioritized by their index in T, lowest index first.

See also schedule_global for more general global scheduling.

RealTimeScheduling.AbstractSchedule — Type

AbstractSchedule{T}

Abstract supertype for all real-time schedules.

RealTimeScheduling.AbstractTaskSchedule — Type

AbstractTaskSchedule{T}

Abstract supertype for all real-time schedules of task systems.

RealTimeScheduling.RealTimeTaskSchedule — Type

RealTimeTaskSchedule{T}(tasks::AbstractRealTimeTaskSystem, jobs::Vector{Vector{T}})

Schedule of a real-time task system.

RealTimeScheduling.RealTimeTaskSchedule — Method

RealTimeTaskSchedule(T::Type, tasks::AbstractRealTimeTaskSystem)

Create a RealTimeTaskSchedule{T} with an empty job list for each task.

Concrete Jobs

Schedule objects contain a Vector of Vectors of AbstractJob objects. These contain all the parameters of a concrete real-time job: release, deadline, cost, and priority, as well as a Vector of ExecInterval objects.

RealTimeScheduling.AbstractJob — Type

AbstractJob

Abstract supertype for all real-time jobs.

RealTimeScheduling.AbstractJobOfTask — Type

AbstractJobOfTask{T}

Abstract supertype for all real-time jobs of tasks.

RealTimeScheduling.Job — Type

Job{S}(release::S, deadline::S, cost::S, priority::S, exec::Vector{ExecInterval{S}}

A real-time job with the given parameters, executing over the intervals in exec.

RealTimeScheduling.JobOfTask — Type

JobOfTask{S, T}(task::T, release::S, deadline::S, cost::S, priority::S, exec::Vector{ExecInterval{S}}

A real-time job of the given task with the given parameters, executing over the intervals in exec.

RealTimeScheduling.task — Method

task(j::AbstractJobOfTask)

Return the task associated with job j.

RealTimeScheduling.release — Method

release(j::AbstractJob)

Return the release time of job j.

RealTimeScheduling.deadline — Method

deadline(j::AbstractJob)

Return the absolute deadline of job j.

RealTimeScheduling.cost — Method

cost(j::AbstractJob)

Return the execution cost of job j.

RealTimeScheduling.priority — Method

priority(j::AbstractJob)

Return the priority of job j.

RealTimeScheduling.exec — Function

exec(j::Job)

Return the execution intervals of job j.

exec(j::JobOfTask)

Return the execution intervals of job j.

RealTimeScheduling.ExecInterval — Type

ExecInterval{S}(start::S, stop::S, proc::Int)

Interval type for job execution, with a specified processor index. Always assumed to be half open, i.e., [start, stop).